pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shift-calibration-lab"
version = "0.1.0"
description = "Calibration and distribution-shift experimentation framework"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Priyojit Palit" }
]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "matplotlib>=3.7",
  "PyYAML>=6.0"
]

[project.scripts]
scalab = "scalab.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
